pyproject.toml
[tool.poetry]
name = "rememberme"
version = "0.1.0"
description = ""
authors = ["Dima Danyliuk <dimadanylyuk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
django-extensions = "^3.1.1"
python-dotenv = "^0.15.0"
psycopg2 = "^2.8.6"
pycountry = "^20.7.3"
translate = "^3.5.0"
python-datamuse = "^1.3.0"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
